Remove 'recurse' parameter from Inline.
The way it is implemented currently is bad because while we're inlining
to a graph G, we are also mutating all the graphs that are being
inlined. The problem is that the graphs we're inlining are usually the
original graphs of functions, so we're silently changing them behind the
scenes, and we don't have a way to recover 'unoptimized' graphs
afterwards.

ghstack-source-id: ccb4f5f92456b7a7472e0adbd8ab932adb9a3e5a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/26487