Use optimized graph in Inline (essentially, making Inline recursive now).
This basically fixes Inline(recurse=true) and makes it a default. One
reservation against running inlining recursively in the original
implementation was that we might hit a quadratic behavior, but in this
implementation it's not an issue since we're inlining only already
inlined graphs and as we recursively descend the call tree we're caching
graphs we've already optimized.

ghstack-source-id: f9ce87bfef3715e89bdac7f3a242a47c0dbd30ce
Pull Request resolved: https://github.com/pytorch/pytorch/pull/26489